mmocr/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-multipart==0.0.6
redis==4.6.0
pillow
numpy < 2.0.0
opencv-python-headless
onnxruntime